+++ backend-java/DZShop/product-service/src/main/java/be/pxl/services/controller/CategoryController.java
@@ -2,6 +2,8 @@ package be.pxl.services.controller;
 
 import be.pxl.services.domain.*;
 import be.pxl.services.services.CategoryService;
+import be.pxl.services.services.HeaderValidationService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,12 +11,15 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Base64;
+
 @RestController
 @RequestMapping("/category")
 @RequiredArgsConstructor
 public class CategoryController {
     private static final Logger log = LoggerFactory.getLogger(CategoryController.class);
     private final CategoryService categoryService;
+    private final HeaderValidationService headerValidationService;
 
     @GetMapping
     public ResponseEntity getAllCategories() {
@@ -23,30 +28,53 @@ public class CategoryController {
     }
 
     @PostMapping
-    public ResponseEntity createCategory(@RequestBody CategoryRequest categoryRequest) {
-        log.info("Create category endpoint called with body {}", categoryRequest);
-        return new ResponseEntity(categoryService.addCategory(categoryRequest), HttpStatus.CREATED);
+    public ResponseEntity createCategory(@RequestHeader(value = "Authorization", required = false) String authorizationHeader ,@RequestBody CategoryRequest categoryRequest) {
+        boolean isAdmin = headerValidationService.hasAdminRole(authorizationHeader);
+        if (isAdmin) {
+            log.info("Create category endpoint called with body {}", categoryRequest);
+            return new ResponseEntity(categoryService.addCategory(categoryRequest), HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity(HttpStatus.FORBIDDEN);
+        }
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity updateCategory(@PathVariable Long id, @RequestBody CategoryRequest categoryRequest) {
-        log.info("Update category endpoint called with categoryId {} and body {}", id, categoryRequest);
-        return new ResponseEntity(categoryService.updateCategory(id,categoryRequest), HttpStatus.OK);
+    public ResponseEntity updateCategory(@RequestHeader(value = "Authorization", required = false) String authorizationHeader ,@PathVariable Long id, @RequestBody CategoryRequest categoryRequest) {
+        boolean isAdmin = headerValidationService.hasAdminRole(authorizationHeader);
+        if (isAdmin) {
+            log.info("Update category endpoint called with categoryId {} and body {}", id, categoryRequest);
+            return new ResponseEntity(categoryService.updateCategory(id,categoryRequest), HttpStatus.OK);
+        } else {
+            return new ResponseEntity(HttpStatus.FORBIDDEN);
+        }
     }
 
     @PostMapping("/{categoryId}/products/{productId}")
     public ResponseEntity addProductToCategory(
+            @RequestHeader(value = "Authorization", required = false) String authorizationHeader ,
             @PathVariable Long categoryId,
             @PathVariable Long productId) {
-        log.info("Add product to category endpoint called with categoryId {} and productId {}", categoryId, productId);
-        return ResponseEntity.ok(categoryService.addProductToCategory(categoryId, productId));
+        boolean isAdmin = headerValidationService.hasAdminRole(authorizationHeader);
+        if (isAdmin) {
+            log.info("Add product to category endpoint called with categoryId {} and productId {}", categoryId, productId);
+            return ResponseEntity.ok(categoryService.addProductToCategory(categoryId, productId));
+        } else {
+            return new ResponseEntity(HttpStatus.FORBIDDEN);
+        }
     }
 
     @DeleteMapping("/{categoryId}/products/{productId}")
     public ResponseEntity removeProductFromCategory(
+            @RequestHeader(value = "Authorization", required = false) String authorizationHeader ,
             @PathVariable Long categoryId,
             @PathVariable Long productId) {
-        log.info("Remove product from category endpoint called with categoryId {} and productId {}", categoryId, productId);
-        return ResponseEntity.ok(categoryService.removeProductFromCategory(categoryId,productId));
+        boolean isAdmin = headerValidationService.hasAdminRole(authorizationHeader);
+        if (isAdmin) {
+            log.info("Remove product from category endpoint called with categoryId {} and productId {}", categoryId, productId);
+            return ResponseEntity.ok(categoryService.removeProductFromCategory(categoryId,productId));
+        } else {
+            return new ResponseEntity(HttpStatus.FORBIDDEN);
+        }
     }
+
 }
